Separate files for different functionalities of the tool
